Add assimp submodule, move import files
