+++ tools/editor/io_plugins/SCsub
@@ -2,6 +2,4 @@ Import('env')
 Export('env')
 env.add_source_files(env.tool_sources,"*.cpp")
 
-
-
-
+SConscript('assimp/SCsub')
