+++ tools/editor/io_plugins/assimp/editor_import_collada.cpp
